+++ pkg/plugin/twinmaker/handler.go
@@ -36,34 +36,6 @@ type twinMakerHandler struct {
 	client TwinMakerClient
 }
 
-type alarm struct {
-	time       *time.Time
-	name       *string
-	status     *string
-	id         *string
-	entityId   *string
-	entityName *string
-}
-
-func (a *alarm) sortString() string {
-	t := time.Unix(0, 0)
-	if a.time != nil {
-		t = *a.time
-	}
-
-	e := "ent"
-	if a.entityName != nil {
-		e = *a.entityName
-	}
-
-	n := "__"
-	if a.name != nil {
-		n = *a.name
-	}
-
-	return fmt.Sprintf("%v/%v/%v", t, e, n)
-}
-
 func NewTwinMakerHandler(client TwinMakerClient) TwinMakerHandler {
 	return &twinMakerHandler{
 		client: client,
@@ -76,6 +48,12 @@ func (s *twinMakerHandler) ListWorkspaces(ctx context.Context, query models.Twin
 	if err != nil {
 		return
 	}
+
+	if results == nil {
+		dr.Error = fmt.Errorf("error loading workspaces")
+		return
+	}
+
 	fields := newTwinMakerFrameBuilder(len(results.WorkspaceSummaries))
 
 	arn := fields.ARN()
@@ -101,6 +79,12 @@ func (s *twinMakerHandler) ListScenes(ctx context.Context, query models.TwinMake
 	if err != nil {
 		return
 	}
+
+	if results == nil {
+		dr.Error = fmt.Errorf("error loading scenes")
+		return
+	}
+
 	fields := newTwinMakerFrameBuilder(len(results.SceneSummaries))
 
 	arn := fields.ARN()
@@ -126,6 +110,12 @@ func (s *twinMakerHandler) ListEntities(ctx context.Context, query models.TwinMa
 	if err != nil {
 		return
 	}
+
+	if results == nil {
+		dr.Error = fmt.Errorf("error loading entities")
+		return
+	}
+
 	fields := newTwinMakerFrameBuilder(len(results.EntitySummaries))
 
 	entityId := fields.EntityID()
@@ -153,6 +143,12 @@ func (s *twinMakerHandler) ListComponentTypes(ctx context.Context, query models.
 	if err != nil {
 		return
 	}
+
+	if results == nil {
+		dr.Error = fmt.Errorf("error loading componentTypes")
+		return
+	}
+
 	fields := newTwinMakerFrameBuilder(len(results.ComponentTypeSummaries))
 
 	componentId := fields.ComponentID()
@@ -176,6 +172,11 @@ func (s *twinMakerHandler) GetEntity(ctx context.Context, query models.TwinMaker
 	result, err := s.client.GetEntity(ctx, query)
 	dr.Error = err
 	if err != nil {
+		return
+	}
+
+	if result == nil {
+		dr.Error = fmt.Errorf("error loading entity")
 		return
 	}
 
@@ -235,6 +236,11 @@ func (s *twinMakerHandler) GetPropertyValue(ctx context.Context, query models.Tw
 	results, err := s.client.GetPropertyValue(ctx, query)
 	dr.Error = err
 	if err != nil {
+		return
+	}
+
+	if results == nil {
+		dr.Error = fmt.Errorf("error loading propertyValue")
 		return
 	}
 
@@ -325,9 +331,14 @@ func (s *twinMakerHandler) processMapValue(v map[string]*iottwinmaker.DataValue)
 	return frame
 }
 
-func (s *twinMakerHandler) processHistory(results *iottwinmaker.GetPropertyValueHistoryOutput, err error, query models.TwinMakerQuery) (dr backend.DataResponse) {
+func (s *twinMakerHandler) processHistory(results *iottwinmaker.GetPropertyValueHistoryOutput, err error, failures []data.Notice, query models.TwinMakerQuery) (dr backend.DataResponse) {
 	dr.Error = err
 	if err != nil {
+		return
+	}
+
+	if results == nil {
+		dr.Error = fmt.Errorf("error loading propertyValueHistory")
 		return
 	}
 
@@ -336,12 +347,17 @@ func (s *twinMakerHandler) processHistory(results *iottwinmaker.GetPropertyValue
 			continue
 		}
 		fields := newTwinMakerFrameBuilder(len(prop.Values))
-		t := fields.Time()
+		// Must return value field first so its labels can be used for the Time field
 		v, conv := fields.Value(prop.Values[0].Value)
+		t := fields.Time()
 		v.Name = "" // filled in with value below
 		for i, history := range prop.Values {
-			t.Set(i, history.Timestamp)
-			v.Set(i, conv(history.Value))
+			if timeValue, err := getTimeObjectFromStringTime(history.Time); err == nil {
+				t.Set(i, timeValue)
+				v.Set(i, conv(history.Value))
+			} else {
+				dr.Error = fmt.Errorf("error parsing timestamp while loading propertyValueHistory")
+			}
 		}
 
 		ref := prop.EntityPropertyReference
@@ -366,20 +382,35 @@ func (s *twinMakerHandler) processHistory(results *iottwinmaker.GetPropertyValue
 		}
 
 		frame := fields.ToFrame("", results.NextToken)
+		frame.AppendNotices(failures...)
 		dr.Frames = append(dr.Frames, frame)
 	}
 	return
 }
 
-func (s *twinMakerHandler) GetComponentHistory(ctx context.Context, query models.TwinMakerQuery) backend.DataResponse {
+func (s *twinMakerHandler) GetComponentHistory(ctx context.Context, query models.TwinMakerQuery) (dr backend.DataResponse) {
 	if query.ComponentTypeId == "" {
 		return backend.DataResponse{
 			Error: fmt.Errorf("missing component parameter"),
 		}
 	}
 
-	result, err := s.client.GetPropertyValueHistory(ctx, query)
-	return s.processHistory(result, err, query)
+	propertyReferences, failures, err := s.GetComponentHistoryWithLookup(ctx, query)
+	result := &iottwinmaker.GetPropertyValueHistoryOutput{
+		NextToken:      nil,
+		PropertyValues: []*iottwinmaker.PropertyValueHistory{},
+	}
+
+	for _, p := range propertyReferences {
+		propertyValue := iottwinmaker.PropertyValueHistory{
+			EntityPropertyReference: p.entityPropertyReference,
+			Values:                  p.values,
+		}
+		result.PropertyValues = append(result.PropertyValues, &propertyValue)
+	}
+
+	// Return dataFrame with the history results and entityId and componentName
+	return s.processHistory(result, err, failures, query)
 }
 
 func (s *twinMakerHandler) GetEntityHistory(ctx context.Context, query models.TwinMakerQuery) backend.DataResponse {
@@ -389,11 +420,13 @@ func (s *twinMakerHandler) GetEntityHistory(ctx context.Context, query models.Tw
 		}
 	}
 	result, err := s.client.GetPropertyValueHistory(ctx, query)
-	return s.processHistory(result, err, query)
+	failures := []data.Notice{}
+	return s.processHistory(result, err, failures, query)
 }
 
-// return status and value here
+// Variation of GetComponentHistory for all alarm components that extend from the basic componentType
 func (s *twinMakerHandler) GetAlarms(ctx context.Context, query models.TwinMakerQuery) (dr backend.DataResponse) {
+	failures := []data.Notice{}
 	alarmComponentType := "com.amazon.iottwinmaker.alarm.basic"
 	externalIdKey := "alarm_key"
 	alarmProperty := "alarm_status"
@@ -412,37 +445,37 @@ func (s *twinMakerHandler) GetAlarms(ctx context.Context, query models.TwinMaker
 	}
 
 	// Get all componentTypes that extend from the base alarm type
-	alarmComponentTypes := map[string]*iottwinmaker.ComponentTypeSummary{}
 	query.ComponentTypeId = alarmComponentType
 	componentTypes, err := s.client.ListComponentTypes(ctx, query)
 	dr.Error = err
 	if err != nil {
 		return
 	}
+	if componentTypes == nil {
+		dr.Error = fmt.Errorf("error loading componentTypes for GetAlarms query")
+		return
+	}
 
 	// Get the propertyValueHistory associated with all componentTypes from above
-	var pValues []*iottwinmaker.PropertyValueHistory
+	var pValues []PropertyReference
+
 	for _, componentTypeSummary := range componentTypes.ComponentTypeSummaries {
 		// Set mapping of alarm component types for quick lookup later
-		alarmComponentTypes[*componentTypeSummary.ComponentTypeId] = componentTypeSummary
 		query.EntityId = ""
 		query.Properties = []*string{aws.String(alarmProperty)}
 		query.ComponentTypeId = *componentTypeSummary.ComponentTypeId
+		query.Order = models.ResultOrderDesc
 		if isFiltered {
 			query.PropertyFilter = filter
 		}
-		p, err := s.client.GetPropertyValueHistory(ctx, query)
+
+		propertyReferences, newFailures, err := s.GetComponentHistoryWithLookup(ctx, query)
 		dr.Error = err
 		if err != nil {
 			return
 		}
-		pValues = append(pValues, p.PropertyValues...)
-		if isLimited && len(pValues) >= maxResults {
-			if len(pValues) > maxResults {
-				pValues = pValues[:maxResults]
-			}
-			break
-		}
+		failures = append(failures, newFailures...)
+		pValues = append(pValues, propertyReferences...)
 	}
 
 	fields := newTwinMakerFrameBuilder(len(pValues))
@@ -486,61 +519,20 @@ func (s *twinMakerHandler) GetAlarms(ctx context.Context, query models.TwinMaker
 	}
 	t := fields.Time()
 
-	query.EntityId = ""
-	query.Properties = nil
-	query.ComponentTypeId = ""
-	failures := []data.Notice{}
-	for i, alarm := range pValues {
-		externalId := alarm.EntityPropertyReference.ExternalIdProperty[externalIdKey]
-		query.ListEntitiesFilter = []models.TwinMakerListListEntitiesFilter{
-			{
-				ExternalId: *externalId,
-			},
-		}
-		le, err := s.client.ListEntities(ctx, query)
-
-		if err != nil {
-			notice := data.Notice{
-				Severity: data.NoticeSeverityWarning,
-				Text:     err.Error(),
+	for i, propertyReference := range pValues {
+		aValues := len(propertyReference.values)
+		if aValues > 0 {
+			if timeValue, err := getTimeObjectFromStringTime(propertyReference.values[0].Time); err == nil {
+				t.Set(i, timeValue)
+			} else {
+				dr.Error = fmt.Errorf("error parsing timestamp during GetAlarms query")
 			}
-			failures = append(failures, notice)
-			break
+			status.Set(i, propertyReference.values[0].Value.StringValue)
 		}
-		entityId := le.EntitySummaries[0].EntityId
-		entityName := le.EntitySummaries[0].EntityName
-		query.EntityId = *entityId
-		e, err := s.client.GetEntity(ctx, query)
-		if err != nil {
-			notice := data.Notice{
-				Severity: data.NoticeSeverityWarning,
-				Text:     err.Error(),
-			}
-			failures = append(failures, notice)
-			break
-		}
-		componentName := ""
-		for _, component := range e.Components {
-			_, isAlarm := alarmComponentTypes[*component.ComponentTypeId]
-			if isAlarm {
-				for propertyKey, propertyValue := range component.Properties {
-					if propertyKey == "alarm_key" {
-						if *propertyValue.Value.StringValue == *externalId {
-							componentName = *component.ComponentName
-							break
-						}
-					}
-				}
-				break
-			}
-		}
-		aValues := len(alarm.Values)
-		t.Set(i, alarm.Values[aValues-1].Timestamp)
-		name.Set(i, &componentName)
-		status.Set(i, alarm.Values[aValues-1].Value.StringValue)
-		id.Set(i, externalId)
-		eId.Set(i, entityId)
-		eName.Set(i, entityName)
+		name.Set(i, propertyReference.entityPropertyReference.ComponentName)
+		id.Set(i, propertyReference.entityPropertyReference.ExternalIdProperty[externalIdKey])
+		eId.Set(i, propertyReference.entityPropertyReference.EntityId)
+		eName.Set(i, propertyReference.entityName)
 	}
 	frame := fields.ToFrame("", nil)
 	frame.AppendNotices(failures...)