+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.1.0
+
+- Upgrade Go AWS SDK.
+- Update aws-iot-twinmaker-grafana-utils package.
+- Fix minor bugs with the datasource configuration page.
+
 ## v1.0.1
 
 - Update aws-iot-twinmaker-grafana-utils package.